+++ src/bin/todor/main.rs
@@ -109,16 +109,19 @@ fn run(matches: &ArgMatches) -> Result<i32, Error> {
         }
     }
 
+    let check = matches.is_present("CHECK");
     if matches.is_present("DELETE_MODE") {
         run_delete(&mut todor)?;
     } else if let Some(p) = pred {
         todor.print_filtered_todos(&p);
+        if check && todor.num_filtered_todos(&p) > 0 {
+            return Ok(1);
+        }
     } else {
         todor.print_todos();
-    }
-
-    if matches.is_present("CHECK") && todor.num_todos() > 0 {
-        return Ok(1);
+        if check && todor.num_todos() > 0 {
+            return Ok(1);
+        }
     }
 
     Ok(0)