+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - Fixed output when tags are styled to be underlined
+- Fixed `--check` output when you use `--user` to filter output
 
 
 ## v0.5.0 (2018-12-16)